Collapse all routing logic into Router
