+++ test/cachex/services_test.exs
@@ -22,7 +22,30 @@ defmodule Cachex.ServicesTest do
              },
              %{id: Services.Informant, start: {Services.Informant, _, _}},
              %{id: Services.Incubator, start: {Services.Incubator, _, _}},
-             %{id: Services.Conductor, start: {Services.Conductor, _, _}},
+             %{id: Services.Courier, start: {Services.Courier, _, _}},
+             %{id: Services.Janitor, start: {Services.Janitor, _, _}}
+           ] = Services.cache_spec(cache)
+  end
+
+  test "generating cache specifications with routing" do
+    # generate the test cache state using an async router
+    name = TestUtils.create_cache(router: Cachex.Router.Ring)
+    cache = Services.Overseer.retrieve(name)
+
+    # validate the services
+    assert [
+             %{id: Eternal, start: {Eternal, _, _}},
+             %{id: ExHashRing.Ring, start: {ExHashRing.Ring, _, _}},
+             %{
+               id: Cachex.Router.Ring.Monitor,
+               start: {GenServer, _, _}
+             },
+             %{
+               id: Services.Locksmith.Queue,
+               start: {Services.Locksmith.Queue, _, _}
+             },
+             %{id: Services.Informant, start: {Services.Informant, _, _}},
+             %{id: Services.Incubator, start: {Services.Incubator, _, _}},
              %{id: Services.Courier, start: {Services.Courier, _, _}},
              %{id: Services.Janitor, start: {Services.Janitor, _, _}}
            ] = Services.cache_spec(cache)
@@ -42,7 +65,6 @@ defmodule Cachex.ServicesTest do
              },
              %{id: Services.Informant, start: {Services.Informant, _, _}},
              %{id: Services.Incubator, start: {Services.Incubator, _, _}},
-             %{id: Services.Conductor, start: {Services.Conductor, _, _}},
              %{id: Services.Courier, start: {Services.Courier, _, _}}
            ] = Services.cache_spec(cache)
   end
